Revert "Use int for unix timestamp"
This reverts commit 94120c476fff49dbac85fbab7301c20ddb32909a.

+++ Mixpanel/Mixpanel.cs
@@ -132,10 +132,10 @@ namespace mixpanel
             return "none";
         }
 
-        private static int CurrentTime()
+        private static double CurrentTime()
         {
             DateTime epochStart = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
-            int currentEpochTime = (int)(DateTime.UtcNow - epochStart).TotalSeconds;
+            double currentEpochTime = (DateTime.UtcNow - epochStart).TotalSeconds;
             return currentEpochTime;
         }
